test: move minicpmo tests to subdirectory with shared 3-stage conftest, fix num_cards and event types
