+++ tests/e2e/online_serving/minicpmo/run_minicpmo_realtime_duplex_demo_pair.py
